+++ api/api.js
@@ -1,14 +1,5 @@
 import axios from "axios";
 
-export const getCurriculums = async () => {
-	try {
-		const response = await axios.get("/api/curriculums");
-		return response.data;
-	} catch (error) {
-		console.error(error);
-	}
-};
-
 export const postResource = async (data) => {
 	try {
 		const response = await axios.post(
