+++ pages/home.js
@@ -1,91 +1,39 @@
 import React, { useState, useEffect } from "react";
-
+import styles from "../styles/Home.module.css";
 import { useRouter } from "next/router";
-import { postResource } from "../api/api";
-
 import BasicCard from "../components/BasicCard/BasicCard";
 import axios from "axios";
 
-// MUI
-import Button from "@mui/material/Button";
-import Alert from "@mui/material/Alert";
-import Stack from "@mui/material/Stack";
-
 function Home() {
-  const router = useRouter();
-  const [data, setData] = useState([]);
-  const [showAlert, setShowAlert] = useState(false);
+	const router = useRouter();
+	const [data, setData] = useState([]);
 
-  useEffect(() => {
-    const getCurriculums = async () => {
-      try {
-        const response = await axios.get("/api/curriculums");
-        setData(response.data);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-    getCurriculums();
-  }, []);
+	useEffect(() => {
+		const getCurriculums = async () => {
+			try {
+				const response = await axios.get("/api/curriculums");
+				setData(response.data);
+			} catch (error) {
+				console.error(error);
+			}
+		};
+		getCurriculums();
+	}, []);
 
-  const submitPost = async () => {
-    //TODO manage real data
-    postResource({
-      curriculumId: "4",
-      learningUnitId: "2",
-      name: "Hola mundo",
-      url: "https://css-tricks.com/",
-      description: "Un gran curso de css",
-    });
-  };
-
-  return (
-    <div>
-      <h1>Hola Usuario</h1>
-      <div>
-        {showAlert && (
-          <Stack sx={{ width: "100%" }} spacing={2}>
-            <Alert severity="error">No tienes una sesión activa</Alert>
-          </Stack>
-        )}
-      </div>
-
-      <div>
-        {data?.map((item, index) => (
-          <BasicCard
-            key={index}
-            title={item.name}
-            description={item.description}
-          />
-        ))}
-      </div>
-      <div style={{ margin: "10%" }}>
-        <Button onClick={submitPost} variant="contained">
-          Crear un recurso
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-	const handleRedirect = (currentRoadmap) => {
+	const dataHandler = (currentRoadmap) => {
 		router.push(
 			{
-				pathname: "/roadmaps",
+				pathname: "/roadmap",
 				query: {
 					...currentRoadmap,
 				},
 			},
-			"/roadmaps"
+			"/roadmap"
 		);
 	};
 
 	return (
 		<div className={styles.container}>
-			<Head>
-				<title> 🔥 Paraffin 🔥 </title>
-				<meta name="description" content="Generated by create next app" />
-			</Head>
 			<main className={styles.main}>
 				<h1 className={styles.title}>🔥 Paraffin 🔥</h1>
 				<br></br>
@@ -97,13 +45,14 @@ function Home() {
 					</h2>
 				</div>
 				<div>
-					{data?.map((item, i) => (
-						<p className={styles.description} key={item.id}>
+					{data?.map((roadmap, index) => (
+						<p className={styles.description} key={roadmap.id}>
 							<BasicCard
-								title={item.name}
-								description={item.description}
+								key={index}
+								title={roadmap.name}
+								description={roadmap.description}
 								type="Roadmap"
-								action={() => handleRedirect(item)}
+								action={() => dataHandler(roadmap)}
 							/>
 						</p>
 					))}
@@ -112,3 +61,5 @@ function Home() {
 		</div>
 	);
 }
+
+export default Home;
