+++ pages/resource.js
@@ -1,0 +1,9 @@
+// const submitPost = async () => {
+// 	postResource({
+// 		curriculumId: "4",
+// 		learningUnitId: "2",
+// 		name: "Hola mundo",
+// 		url: "https://css-tricks.com/",
+// 		description: "Un gran curso de css",
+// 	});
+// };
